chore: reorg l1d related files
